week25
improved logic for calculating time between pass fail, changed the index search to find the latest pass and compare it against the latest fail.

If for whatever reason there are many passes afetr the last fail, this logic would break. Need to implement, if there is a pass, find the latest fail and find the next pass right after